Add files via upload
A member of a truss is the simplest solid element, namely, an elastic rod with ends 1 and 2 referred to hereafter as nodes. Consider an element of length L, cross-section A, and made of a linear elastic material having Young’s modulus E. If we apply a normal force N2 at node 1, and at the same time maintain nodes 2 and 3 fixed.